Correct a bug found in the TIMING routines on an ALPHA platform.    --> It crashed on one TIMING program, zsvdtim.in
Comment below made by Edward Smyth:
-----------------------------------
I believe the problem with dlasdq when timing using the zsvdtim.in input file is due to a bug in TIMING/EIG/ztimee.f. In the call to ZTIM26, IWORK is passed as an actual argument instead of the much larger IWORK2, which is used in the ZTIM21 and ZTIM22 calls, and used in the call to DTIM26 in dtimee.f. IWORK is too small for the requirements in ZTIM26. With this corrected, it worked OK on our Alpha system.

ctimee.f has the same bug, but didn''t actually fail. Also, I didn''t have any problems with ctime.in.

This was tested on an EV6 with source code BLAS, and using Compaq Fortran compiler V5.4A, using flags f90 -O3 -fpe0 to compile. 

-----------------------------------------------------------------------------------------------------------------------------------
See http://icl.cs.utk.edu/lapack-forum/viewtopic.php?t=78 for more information