Rename new xlahrd to xlahr2, restore old xlahrd.
Follow up on the previous commit for Hessenberg by Daniel Kressner.

As the new xlahrd.f is not backward compatible (see previous log), we
have to keep the 'old' xlahrd and add the new xlahrd routine as xlahr2.

Add the following comment in xlahrd routines:
*  This is an OBSOLETE auxiliary routine.
*  This routine will be 'deprecated' in a  future release.
*  Please use the new routine DLAHR2 instead.

Change call in xgehrd so that xlahr2 is called instead of xlahrd.

Modify the SRC/Makefile to add these new routines.