Add ILA_LEN_TRIM to prepare for longer names.
ILA_LEN_TRIM returns the integer offset of the last
non-blank character in its argument, or 0 if there
is none.  It's a portable F77 version of the common
LEN_TRIM intrinsic.  It could be implemented using
the modern intrinsic VERIFY, as well.

Slicing S(1:ILA_LEN_TRIM(S)) in WRITE statements
hopefully will not generate a copy.  The F90 TRIM
intrinsic isn't clear on when it copies, either.

The routine is added to both LAPACK and the reference
BLAS; they both provide a XERBLA which will need it.

Signed-off-by: Jason Riedy <ejr@cs.berkeley.edu>