Move updated xLAMCH routines to INSTALL/
The xLAMCH routines are copied from INSTALL/ to SRC/ on every
top-level make.  Having two copies can be confusing.  In this
case, the ones in SRC/ were updated without changing the ones
in INSTALL/, so the current wrapper generator would mis-parse
the source if run *after* the library is built.

Signed-off-by: Jason Riedy <ejr@cs.berkeley.edu>