Apply a patch from Bobby Cheng, MathWorks for unblocked LDLT routines.
Notes from conversation between Sven, Bobby and Jason:

The problem is the behavior of MAX with respect to different compilers. 

*-------------------------------------------------
| g77 and g95 MAX       |  gfortran MAX         |
*-------------------------------------------------
| function x = max(a,b) | function x = max(a,b) |
| if a > b              | if a < b              |
|    x = a              |    x = b              |
| else                  | else                  |
|    x = b              |    x = a              |
| end                   | end                   |
*-------------------------------------------------

Note that both implementation are valid with the current Fortran standard.

Consequently g77 and g95 MAX are such that
   max(NaN, 0) returns 0
   max(0, NaN) returns NaN
In contrast, gfortran implementation behaves like
   max(NaN, 0) returns NaN
   max(0, NaN) returns 0

The former xSY/HETF2 was ok with a g77/g95 implementation but not ok with gfortran.
In the case N = 1 and A(1,1) = NaN, the former LAPACK code with gfortran, tries
to pick a 2 by 2 pivot and hence gets a memory violation.

The patch is the following:

Replace l.211 and l.383
         IF( MAX( ABSAKK, COLMAX ).EQ.ZERO ) THEN

by

         IF( (MAX( ABSAKK, COLMAX ).EQ.ZERO) .OR.
     $       (ABSAKK .NE. ABSAKK) ) THEN

To conclude the log, here is a related remark from Jason:

> Anyone know how heavily performance depends on a fast MAX
> or MIN?  Should we provide our own, symmetric, number-
> preferring xLAMAX and xLAMIN?  Then we could "simply"
> replace all the MAXs and MINs with those functions.