Import single-precision MRRR updates. From: Christof Vömel <cvoemel@lbl.gov>
To keep with the six-character limit in the 3.1 release, the new MRRR
routine is named SSTEMR rather than the previously agreed SSTEMR3.

The SSTEGR interface ignores its ABSTOL argument and calls the new
SSTEMR.  SSTEGR also promises that there are N columns in the
eigenvector array Z.  SSTEGR itself does not have enough information
to know how many columns are necessary.

All new routines include documentation.  The SSTEGR documentation is
updated.

The new SSTEMR routine uses INFO as an input/output argument.  This
will be changed in an upcomming commit.

NOTE: The following helper routines have a different interface but the
same name:
  - SLAR1V
  - SLARRB
  - SLARRE
  - SLARRF
  - SLARRV
Unlike helpers like xLACON, each of these is very specific to the
overall algorithm.  The interface can be changed because there is
little possibility of outside use.

Acked-by: Jason Riedy <ejr@cs.berkeley.edu>