The ISNAN protection added in the complex version of SYTF2 was messed up.
ABSAKK is a REAL or DOUBLE PRECISION number (as his name indicated, it
represents the absolute value of a COMPLEX or COMPLEX*16).

So we should use:
either DISNAN( ABSAKK )
or     ZISNAN( A( K, K ) )
I have chosen the first option.

So the two statements:
      IF( MAX( ABSAKK, COLMAX ).EQ.ZERO .OR. ZISNAN(ABSAKK) ) THEN
are replaced by
      IF( MAX( ABSAKK, COLMAX ).EQ.ZERO .OR. DISNAN(ABSAKK) ) THEN
in ZSYTF2.

Same fix in CSYTF2.
Thanks for Julie to point this out and to GCC-3.4.4 to nicely fails on this.
Everything runs smoothly now.